+++ _editor/assets/elements.js
@@ -1158,7 +1158,7 @@
 
 				const select = dc.createElement('select');
 				select.id = 'elements-form-locale-select';
-				select.className = 'nino-admin-contextbar-select';
+				select.className = 'nino-admin-locale-select nino-admin-contextbar-select';
 				Nino.editor.elements._locales.forEach( function( locale ) {
 					const option = dc.createElement('option');
 					option.value = locale;