+++ _editor/assets/text.js
@@ -483,7 +483,7 @@
 
 				const select = dc.createElement('select');
 				select.id = 'text-form-locale-select';
-				select.className = 'nino-admin-contextbar-select';
+				select.className = 'nino-admin-locale-select nino-admin-contextbar-select';
 				Nino.editor.text._locales.forEach( function( locale ) {
 					const option = dc.createElement('option');
 					option.value = locale;